chore: fix fmt and clippy lints for newer stable Rust
Resolves CI core failures on main after the stable toolchain tightened
several lints (needless_lifetimes via is_multiple_of, collapsible_if,
field_reassign_with_default, unnecessary_cast, unnecessary_sort_by,
inconsistent_digit_grouping) and surfaced pre-existing unused imports
and dead code across the workspace.

+++ crates/engine/ndn-security/src/cert_fetcher.rs
@@ -124,7 +124,7 @@ mod tests {
         // Name
         signed.write_nested(0x07, |w| {
             for comp in name.components() {
-                w.write_tlv(comp.typ as u64, &comp.value);
+                w.write_tlv(comp.typ, &comp.value);
             }
         });
         // Content with public key
